requirements.txt
pymupdf
openai
python-dotenv
Pillow
reportlab
Flask
flask-cors
